consolidate the karma config to tools/karma.conf.js
